+++ functorch/experimental/control_flow.py
@@ -1,1 +1,2 @@
 from ._map import map  # noqa: F401
+from ._cond import cond  # noqa: F401
